Add LOCK file

+++ .github/workflows/test_compose.sh
@@ -3,17 +3,19 @@
 set +e
 set -x
 
-
+if [ -f "./DOCKER_COMPOSE_TEST_LOCK" ]; then
+    echo "A test is already running. Exiting."
+    exit 1
+fi
+touch "./DOCKER_COMPOSE_TEST_LOCK"
 
 GIT_BRANCH="$1"
 VERSION=v$(cat compose.yml | grep 'image: counterparty/counterparty:' | awk -F ":" '{print $3}')
-
 
 # pull the latest code
 rm -rf counterparty-core
 git clone --branch "$GIT_BRANCH" https://github.com/CounterpartyXCP/counterparty-core.git
 cd counterparty-core
-git checkout $GIT_BRANCH
 
 # stop the running containers
 docker compose stop
@@ -43,3 +45,5 @@ curl -X POST http://127.0.0.1:14000/api/ \
      -H 'Content-Type: application/json; charset=UTF-8'\
      -H 'Accept: application/json, text/javascript' \
      --data-binary '{ "jsonrpc": "2.0", "id": 0, "method": "get_running_info" }'
+
+rm -f ../DOCKER_COMPOSE_TEST_LOCK
